fix DB and Services

+++ src/main/java/com/example/end/repository/ProcedureRepository.java
@@ -4,6 +4,6 @@ package com.example.end.repository;
 import com.example.end.models.Procedure;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProcedureRepository extends JpaRepository<Procedure, Long> {
+public interface ProcedureRepository extends JpaRepository<Procedure, Integer> {
 
 }
